+++ libk/memset.c
@@ -1,8 +1,8 @@
-void kmemset(void *const buffer, const unsigned char value, const unsigned int size)
+void kmemset(void *const buffer, const unsigned char value, const unsigned long size)
 {
     unsigned char *const s = buffer;
 
-    for (unsigned int i = 0; i < size; ++i) {
+    for (unsigned long i = 0; i < size; ++i) {
         s[i] = value;
     }
 }
